initial solution for 02-class-id-selectors
Technically this does not complete the challenge, as not every element actually has its own class or id value. initial-solution.png is included to show that it still styles the index.html file to look like the desired solution, but some code does get reused (as seen in the solution).